+++ ui/package.json
@@ -47,7 +47,7 @@
 		"postcss": "^8.4.19",
 		"tailwindcss": "^3.2.4",
 		"typescript": "^4.6.4",
-		"vite": "^3.2.3",
+		"vite": "^4.0.0",
 		"vitest": "^0.25.3"
 	}
 }
