+++ ui/vite.config.ts
@@ -1,6 +1,7 @@
 /// <reference types="vitest" />
 /// <reference types="vite/client" />
 
+import path from 'path';
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 
@@ -14,10 +15,10 @@ export default defineConfig({
 	},
 	resolve: {
 		alias: {
-			'@components': '/src/components',
-			'@pages': '/src/pages',
-			'@features': '/src/features',
-			'@utils': '/src/utils'
+			'@components': path.resolve(__dirname, './src/components'),
+			'@pages': path.resolve(__dirname, './src/pages'),
+			'@features': path.resolve(__dirname, './src/features'),
+			'@utils': path.resolve(__dirname, './src/utils')
 		}
 	}
 });
